1.0

+++ src/main/java/wsg/lol/common/pojo/dto/summoner/SummonerDto.java
@@ -49,7 +49,7 @@ public class SummonerDto extends BaseDto {
     /**
      * Summoner level associated with the summoner.
      */
-    private Long summonerLevel;
+    private Integer summonerLevel;
 
     private Integer score;
